[Finishes #161283706] Rename products views
